CDAP-9148 refactoring to wrap all pipeline plugins
Fixes CDAP-9418 and HYDRATOR-122.
Refactoring so that any known plugin type instantiated by the
pipeline is wrapped so that the context classloader is set
before any plugin method is called, and so that all plugin methods
are timed and a metric emitted for time spent.
Also performs one part of the fix to stage logging, which is to
set a key in the logging MDC before any plugin method is called.